# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DJERDJERA Comptable API - Algerian small-business accounting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List journal entries
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [SALE, PURCHASE, EXPENSE]
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Record a sale, purchase, or expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted

  # ============ INVENTORY ============
...
components:
...
  schemas:
...
    Transaction:
      type: object
      required: [id, date, type, label, amountHt, tvaRate, tvaAmount, amountTtc, paymentMethod, status]
      properties:
        id: { type: string, format: uuid }
        date: { type: string, format: date }
        type: { type: string, enum: [SALE, PURCHASE, EXPENSE] }
        label: { type: string }
        amountHt: { type: number }
        tvaRate: { type: number }
        tvaAmount: { type: number }
        amountTtc: { type: number }
        paymentMethod: { type: string, enum: [CASH, BANK, CREDIT] }
        status: { type: string, enum: [PAID, UNPAID, PARTIAL] }
        thirdParty: { type: string, nullable: true, description: "Customer or supplier name" }
        category: { type: string, nullable: true, description: "e.g. RENT, ELECTRICITY, SUPPLIES, RAW_MATERIALS" }
        itemId: { type: string, format: uuid, nullable: true, description: "Stock article this entry moved, if any" }
        quantity: { type: number, nullable: true, description: "Quantity of the article moved" }
        unitCostHt:
          type: number
          nullable: true
          description: >-
            The CUMP applied when this entry was posted. Frozen at posting time:
            CUMP moves with the stock, so recomputing an old entry at today's
            average would restate books that have already been closed.
        costOfGoodsSold:
          type: number
          nullable: true
          description: "quantity x unitCostHt. Relieved on a sale (class 6); null otherwise."
    TransactionInput:
      type: object
      required: [date, type, label, amountHt, tvaRate, paymentMethod, status]
      properties:
        date: { type: string, format: date }
        type: { type: string, enum: [SALE, PURCHASE, EXPENSE] }
        label: { type: string, minLength: 1 }
        amountHt: { type: number, minimum: 0 }
        tvaRate: { type: number, minimum: 0, maximum: 100 }
        paymentMethod: { type: string, enum: [CASH, BANK, CREDIT] }
        status: { type: string, enum: [PAID, UNPAID, PARTIAL] }
        thirdParty: { type: string, nullable: true }
        category: { type: string, nullable: true }
        itemId:
          type: string
          format: uuid
          nullable: true
          description: >-
            Optional stock article. A SALE naming one relieves inventory at the
            CUMP and records the cost; a PURCHASE naming one feeds it at
            amountHt / quantity. Entries without an itemId behave as before.
        quantity:
          type: number
          nullable: true
          minimum: 0
          description: Required whenever itemId is given.
...